Create Merge Sort
This program implements the Merge Sort algorithm to sort an array of integers in ascending order. It takes user input for the number of elements and the array values, then display the array before and after sorting using the Merge Sort algorithm. Merge Sort is a divide-and-conquer algorithm that recursively divides the array into smaller sub-arrays, sorts them, and then merges them back together to achieve a sorted result.