Create Graph Traversal using DFS and BFS
This C program allows the user to enter details of a graph, represented using an adjacency matrix. It then provides options to perform Depth-First Search (DFS) and Breadth-First Search (BFS) traversals starting from the first vertex (vertex 1). The program displays the vertices visited during the traversals. The user can repeat the process or exit the program.